SolutionSort: Allow the user to specify concurrency
By using the `-j` option, the user can specify fewer threads than what is available, which can restrict the CPU being used. Usually it's best to not use this option and use the defaults. As there is often scanning of files, cores are often not used to full capacity.

Issue: DNX-17